+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Ember Eui Changelog
 
+### 0.0.12
+
+🏠 Internal
+- Let apps define svgPath in environment.js
+  ```ts
+    `ember-eui`: {
+      svgPath: 'svg/assets'
+    }
+  ```
+  This assumes that you setup ember-svg-jar with the prefix `svg/` (which is the default one)
+
+🐛 Bug / Fixes
+- Fixed EuiIcon camelCase svgs
+
 ### 0.0.11
 
 🚀 Enhancements